backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-multipart==0.0.9
pydantic
passlib[bcrypt]==1.7.4
tortoise-orm==0.20.0
aiosqlite
uvicorn[standard]
bcrypt==4.0.1